+++ docs/changelog.rst
@@ -19,6 +19,8 @@ New features:
   control the maximum allowed distance between the lyrics search result and the
   tagged item's artist and title. This is useful for preventing false positives
   when fetching lyrics.
+* :doc:`/plugins/mbsync`: gained a new ``--timid`` flag to print and
+  confirm the changes before applying. :bug:`4250`
 
 Bug fixes: